requirements.txt
amqp==5.0.5
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
autopep8==1.5.6
beautifulsoup4==4.9.3
billiard==3.6.3.0
cached-property==1.5.2
cachetools==4.2.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
defusedxml==0.7.0rc2
dj-rest-auth==2.1.4
Django==3.1.7
django-adminlte-3==0.1.6
django-adminlte3==0.1.6
django-allauth==0.44.0
django-braces==1.14.0
django-celery-beat==2.2.0
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-js-asset==1.2.2
django-oauth-toolkit==1.5.0
django-redis==4.12.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
django-taggit==1.3.0
django-taggit-serializer==0.1.7
django-timezone-field==4.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
inflection==0.5.1
isodate==0.6.0
itypes==1.2.0
Jinja2==2.11.3
jwcrypto==0.8
kombu==5.0.2
lxml==4.6.3
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.9
piexif==1.1.3
Pillow==8.1.1
prompt-toolkit==3.0.17
protobuf==3.15.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
soupsieve==2.2
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
zeep==4.0.0
